[WIP] | Inicio da configuração das rotas no padrão camel

+++ src/main/java/com/repository/FilmeRespository.java
@@ -3,5 +3,5 @@ package com.repository;
 import com.model.Filme;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FilmeRespository extends CrudRepository<Filme, Long> {
+public interface FilmeRespository extends CrudRepository<Filme, Integer> {
 }
